requirements.txt
Flask==3.1.2
Flask-Session==0.8.0
gunicorn==23.0.0
PyPDF2==3.0.1
python-dotenv==1.2.1
stripe==14.0.1
Werkzeug==3.1.3
rank-bm25==0.2.2
psycopg2-binary==2.9.9